pyproject.toml
[tool.poetry]
authors = ["Dan Meshkov <daniil.meshkov@opensea.io>", "Vicky Gong <vicky.gong@opensea.io>", "Taylor Beever <taylor@quicknode.com>"]
description = "Open-Rarity library is an open standard that provides an easy, explanable and reproducible computation for NFT rarity"
license = "Apache-2.0"
name = "openrarity"
version = "1.0.0a0"

readme = "README.md"

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Typing :: Typed",
  "Topic :: Software Development :: Libraries",
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.10",
]
[tool.poetry.scripts]
openrarity = "openrarity.cli.main:app"

[tool.poetry.dependencies]
httpx = "^0.23.0"
pysatchel = "^0.3.1"
python = ">=3.10,<3.12"
tqdm = "^4.64.1"
typer = "^0.6.1"
tabulate = "^0.9.0"
tenacity = "^8.1.0"
numpy = "^1.23.4"
pandas = "^1.5.1"
Cython = "^0.29.32"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.2"
flake8-bugbear = "^22.7.1"
isort = "^5.10.1"
jupyterlab = "^3.4.8"
matplotlib = "^3.6.1"
mypy = "^0.971"
pep8-naming = "^0.13.1"
pre-commit = "^2.19.0"
pytest = "^7.1"
pytest-cov = "^3.0"
types-requests = "^2.28.6"

[tool.black]
line-length = 88

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
include_trailing_comma = true
known_first_party = "openrarity"
multi_line_output = 3
profile = "black"
